moved blog posts to root level so we can split up blog pagination and post logic
