+++ pages/[page].tsx
@@ -1,15 +1,10 @@
-// [slug].js
 import { GetStaticPaths } from 'next';
-import Head from 'next/head';
-import React, { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { Contained } from '../components/Contained';
-import { RichBody } from '../components/RichBody';
+import BlogPost from '../components/BlogPost';
+import RichPage from '../components/RichPage';
 import { CMS, NAVIGATION } from '../constants';
-import { CmsApi, unslugify } from '../services/cms';
-import { PageType, setPageType, SideMenuItem } from '../state/navigation';
-import { ISplitPage } from '../types/cms';
-import { generateTitle, generateURL } from '../utils/metadata';
+import { CmsApi, generateLinkMeta, unslugify } from '../services/cms';
+import { SideMenuItem } from '../state/navigation';
+import { IPost, ISplitPage, isPost } from '../types/cms';
 
 interface IPath {
   params: { page: string };
@@ -20,13 +15,35 @@ export const getStaticPaths: GetStaticPaths = async () => {
   // Hardcoded in navigation constants.
   // Contentful can edit entries but cannot add/remove
   // without touching code.
-  const paths: IPath[] = Object.values(NAVIGATION.SIDE_MENU_ITEMS).map(
-    item => ({
-      params: { page: item.href },
-    }),
-  );
+  const navigationPaths: IPath[] = Object.values(
+    NAVIGATION.SIDE_MENU_ITEMS,
+  ).map(item => ({
+    params: { page: item.href },
+  }));
 
-  return { paths, fallback: 'blocking' };
+  const cms = new CmsApi();
+  let posts: IPost[] = [];
+  let currentPage = 1;
+  let foundAllPosts = false;
+
+  // Contentful only allows 100 at a time
+  while (!foundAllPosts) {
+    const { entries: _posts } = await cms.fetchBlogEntries(100, currentPage);
+
+    if (_posts.length === 0) {
+      foundAllPosts = true;
+      continue;
+    }
+
+    posts = [...posts, ..._posts];
+    currentPage++;
+  }
+
+  const postPaths: IPath[] = posts.map(post => ({
+    params: { page: post.slug },
+  }));
+
+  return { paths: [...navigationPaths, ...postPaths], fallback: 'blocking' };
 };
 
 export async function getStaticProps({ params }) {
@@ -35,7 +52,15 @@ export async function getStaticProps({ params }) {
 
   try {
     const cms = new CmsApi();
-    const page = await cms.fetchPageById(SideMenuItem[id] ?? '');
+    let page: ISplitPage | IPost;
+    if (SideMenuItem[id]) {
+      page = await cms.fetchPageById(SideMenuItem[id] ?? '');
+    } else {
+      page = await cms.fetchEntryBySlug(href, 'post');
+      // embedded links in post body need metadata for preview
+      page.body = await generateLinkMeta(page.body);
+      // TODO might be URL issues
+    }
 
     return {
       props: {
@@ -53,64 +78,12 @@ export async function getStaticProps({ params }) {
   }
 }
 
-function Page({ page, href }: { page: ISplitPage | null; href: string }) {
-  const dispatch = useDispatch();
-  useEffect(() => {
-    dispatch(setPageType(PageType.NORMAL));
-  }, []);
-
-  const pageTitle = generateTitle(page?.label);
-  const pageDescription = page?.title;
-  const pageURL = generateURL(href);
-
-  return (
-    <>
-      <Head>
-        <title>{pageTitle}</title>
-        <meta name="description" content={pageDescription}></meta>
-        <meta property="og:title" content={pageTitle} key="ogtitle" />
-        <meta
-          property="og:description"
-          content={pageDescription}
-          key="ogdesc"
-        />
-        <meta property="og:type" content="website" />
-        <meta
-          property="og:image"
-          content={page?.hero?.imageUrl}
-          key="ogimage"
-        />
-        <meta property="og:url" content={pageURL} />
-
-        <link rel="canonical" href={pageURL}></link>
-        <meta name="twitter:card" content="summary_large_image" />
-        <meta name="twitter:title" content={pageTitle} />
-        <meta name="twitter:description" content={pageDescription} />
-        <meta name="twitter:image" content={page?.hero?.imageUrl} />
-      </Head>
-
-      <div className="bg-alt">
-        <div className="relative flex items-center justify-center w-full h-full pt-3 bg-gradient-to-bl from-hyper to-blush">
-          <img
-            style={{ maxHeight: '33vh' }}
-            src={page?.hero?.imageUrl}
-            className="object-contain w-full"
-            alt={page?.hero?.description ?? pageTitle}
-          />
-        </div>
-
-        <Contained>
-          <h1 className="mt-12 mb-4 text-4xl font-bold leading-none text-primary font-prompt">
-            {page?.title}
-          </h1>
-
-          <div className="mb-10">
-            <RichBody body={page?.body} />
-          </div>
-        </Contained>
-      </div>
-    </>
-  );
+function Page({ page, href }: { page: ISplitPage | IPost; href: string }) {
+  if (isPost(page)) {
+    return <BlogPost post={page} url={href} />;
+  } else {
+    return <RichPage page={page} href={href} />;
+  }
 }
 
 export default Page;
